converter new

+++ appline/src/main/java/ru/job4j/appline/task06/Converter.java
@@ -11,21 +11,21 @@ public class Converter {
             System.out.println("Выберите единицу измерения: 1 - кг, 2 - гр, 3 - тонн");
             int y = scanner.nextInt();
             System.out.println("введите количество выбранных единиц:");
-            int z = scanner.nextInt();
+            double z = scanner.nextDouble();
             switch (y){
                 case 1:
-                    int resKg = z * 1;
-                    int resGr = z * 1000;
-                    double resT = z/1000.0;
+                    double resKg = z * 1;
+                    double resGr = z * 1000;
+                    double resT = z/1000;
                     System.out.println("Результат:");
                     System.out.println("Килограмм: " + resKg);
                     System.out.println("Грамм: " + resGr);
                     System.out.println("Тонн: " + String.format("%.3f", resT));
                     break;
                 case 2:
-                    double resKg2 = z/1000.0;
-                    int resGr2 = z * 1;
-                    double resT2 = z/1000000.0;
+                    double resKg2 = z/1000;
+                    double resGr2 = z * 1;
+                    double resT2 = z/1000000;
                     System.out.println("Результат:");
                     System.out.println("Килограмм: " + String.format("%.3f", resKg2));
                     System.out.println("Грамм: " + resGr2);
@@ -33,8 +33,8 @@ public class Converter {
                     break;
                 case 3:
                     double resKg3 = z * 1000;
-                    int resGr3 = z * 1000000;
-                    int resT3 = z * 1;
+                    double resGr3 = z * 1000000;
+                    double resT3 = z * 1;
                     System.out.println("Результат:");
                     System.out.println("Килограмм: " + resKg3);
                     System.out.println("Грамм: " + resGr3);
@@ -45,10 +45,10 @@ public class Converter {
             System.out.println("Выберите единицу измерения: 1 - метр, 2 - миля, 3 - ярд, 4 - фут");
             int y = scanner.nextInt();
             System.out.println("введите количество выбранных единиц:");
-            int z = scanner.nextInt();
+            double z = scanner.nextDouble();
             switch (y){
                 case 1:
-                    int resM = z * 1;
+                    double resM = z * 1;
                     double resMil = z / 1609.34;
                     double resYar = z * 1.094;
                     double resFut = z * 3.281;
@@ -60,9 +60,9 @@ public class Converter {
                     break;
                 case 2:
                     double resM2 = z * 1609.34;
-                    int resMil2 = z * 1;
-                    int resYar2 = z * 1760;
-                    int resFut2 = z * 5280;
+                    double resMil2 = z * 1;
+                    double resYar2 = z * 1760;
+                    double resFut2 = z * 5280;
                     System.out.println("Результат:");
                     System.out.println("Метры: " + resM2);
                     System.out.println("Мили: " + resMil2);
@@ -71,9 +71,9 @@ public class Converter {
                     break;
                 case 3:
                     double resM3 = z / 1.094;
-                    double resMil3 = z / 1760.0;
-                    int resYar3 = z * 1;
-                    int resFut3 = z * 3;
+                    double resMil3 = z / 1760;
+                    double resYar3 = z * 1;
+                    double resFut3 = z * 3;
                     System.out.println("Результат:");
                     System.out.println("Метры: " + String.format("%.3f", resM3));
                     System.out.println("Мили: " + String.format("%.3f", resMil3));
@@ -82,9 +82,9 @@ public class Converter {
                     break;
                 case 4:
                     double resM4 = z / 3.281;
-                    double resMil4 = z / 5280.0;
-                    double resYar4 = z / 3.0;
-                    int resFut4 = z * 1;
+                    double resMil4 = z / 5280;
+                    double resYar4 = z / 3;
+                    double resFut4 = z * 1;
                     System.out.println("Результат:");
                     System.out.println("Метры: " + String.format("%.2f", resM4));
                     System.out.println("Мили: " + String.format("%.5f", resMil4));
